build_embeds/RETRO-pytorch/pyproject.toml
[tool.poetry]
name = "retro-pytorch"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.21.1"
einops = ">=0.3"
numpy = "^1.23.2"
sentencepiece = "^0.1.97"
torch = ">=1.6"
tqdm = "^4.64.0"
streamlit = "^1.12.0"
jsonlines = "^3.1.0"
codetiming = "^1.3.0"
autofaiss = "^2.15.1"
sacremoses = "^0.0.53"
st-annotated-text = "^3.0.0"
hydra-core = "^1.2.0"
submitit = "^1.4.5"
lz4 = "^4.0.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pylint = "^2.14.5"
isort = "^5.10.1"
ipython = "^8.4.0"
jupyterlab = "^3.4.5"
pytest = "^7.1.2"

[tool.black]
line-length = 128

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
